% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRI}
\alias{TRI}
\title{This is a toy dataset, which is simply an example of a Total Return Index that can be used in packages requiring xts}
\format{A data frame with 16590 rows and 3 variables:
\describe{
  \item{Date}{Valid date column}
  \item{Country}{Country, by which to spread}
  \item{TRI}{Value to xts}
}}
\usage{
TRI
}
\description{
This is a toy dataset, which is simply an example of a Total Return Index that can be used in packages requiring xts
}
\keyword{datasets}

